move `get_party_shipping_address` to party.py
